Add streaming and complete feature
