+++ src/bytes/mod.rs
@@ -1,6 +1,6 @@
 //! Parsers recognizing bytes streams
 
+#[cfg(feature = "complete")]
 pub mod complete;
+#[cfg(feature = "streaming")]
 pub mod streaming;
-#[cfg(test)]
-mod tests;
